pyproject.toml
[project]
name = "station"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "faker>=39.0.0",
    "fastapi>=0.127.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
